ejercicio completo

+++ src/main/kotlin/numero.kt
@@ -1,5 +1,5 @@
 open class numero {
-    var a:Double = 0.0
-    var b: Double = 0.0
-    var c: Double = 0.0
+    var a:Int= 0
+    var b: Int = 0
+    var c: Int = 0
 }
